% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_case_when_formula}
\alias{build_case_when_formula}
\title{Formula for dplyr::case_when}
\usage{
build_case_when_formula(func, value, low, high, string)
}
\arguments{
\item{func}{func}

\item{value}{value}

\item{low}{low}

\item{high}{high}

\item{string}{string}
}
\value{
an expression
}
\description{
Formula for dplyr::case_when
}
