% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_val}
\alias{default_val}
\title{Find the default value, if it exists}
\usage{
default_val(x, value)
}
\arguments{
\item{x, value}{values}
}
\description{
test if first value Truthy, and if so use it, else use a back up value
}
