% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_advConditions.R
\name{mod_advConditions_ui}
\alias{mod_advConditions_ui}
\title{User interface for advConditions module}
\usage{
mod_advConditions_ui(id)
}
\arguments{
\item{id}{standard parameter for {shiny modules}.}
}
\value{
a shiny \code{\link[shiny]{tagList}} containing a well panels of input
  widgets
}
\description{
Provides the interface for user to input important characteristics about new
column the wish to create.
}
