% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_launchSNC.R
\name{mod_launchModal_srv}
\alias{mod_launchModal_srv}
\title{launchModal Server Function}
\usage{
mod_launchModal_srv(id, dat)
}
\arguments{
\item{id}{standard parameter for {shiny modules} used to create namespaces
and connect the UI module to the server module.}

\item{dat}{a r data.frame you want users to  derive a new column from and to.}
}
\value{
a list of reactive objects. The first object is the original
  data.frame in a reactive state (named `data()`), which may contain a newly
  derived column if user successfully created derivation. Secondly, the list
  will also contain `expr()` containing the `dplyr::mutate()` expressions used
  to create said column(s).
}
\description{
launchModal Server Function waits for a button click, and then
  launches a large modal containing a savvy user interface to help users
  create new columns on the fly
}
