% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_launchModal.R
\name{mod_launchModal_ui}
\alias{mod_launchModal_ui}
\title{launchModal UI Function}
\usage{
mod_launchModal_ui(id)
}
\arguments{
\item{id}{standard parameter for {shiny modules}.}
}
\value{
a shiny \code{\link[shiny]{tagList}} containing a simple action
  button to launch the `shinyNewColumns` modal
}
\description{
provides the interface to a single button
}
