% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_newCol.R
\name{mod_newCol_ui}
\alias{mod_newCol_ui}
\title{User interface for newCol module}
\usage{
mod_newCol_ui(id)
}
\arguments{
\item{id}{standard parameter for {shiny modules}.}
}
\value{
a shiny \code{\link[shiny]{tagList}} containing a well panel of input
  widgets
}
\description{
Provides the interface for user to imput important characteristics about new
column the wish to create.
}
