% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rangeConditions.R
\name{mod_rangeConditions_srv}
\alias{mod_rangeConditions_srv}
\title{Server logic for rangeConditions module}
\usage{
mod_rangeConditions_srv(id, dat, grp, reference_var, else_group, else_name)
}
\value{
an expression containing the `dplyr::between()` expressions used to
  create new column
}
\description{
Server logic for rangeConditions module
}
