% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rangeConditions.R
\name{mod_rangeConditions_ui}
\alias{mod_rangeConditions_ui}
\title{User interface for rangeConditions module}
\usage{
mod_rangeConditions_ui(id)
}
\arguments{
\item{id}{standard parameter for {shiny modules}.}
}
\value{
a shiny \code{\link[shiny]{tagList}} containing a well panel of input
  widgets
}
\description{
Provides the interface for user to input important characteristics about new
column the wish to create.
}
